refactor: 테스트 코드 패키지 명 변경
