+++ src/test/java/subway/controller/validator/LineValidatorTest.java
@@ -1,7 +1,11 @@
-package subway.controller.exception;
+package subway.controller.validator;
 
 import org.junit.Test;
 
+import subway.controller.exception.DuplicationException;
+import subway.controller.exception.IllegalElementException;
+import subway.controller.exception.NameFormatException;
+import subway.controller.exception.NotExistedElementException;
 import subway.controller.validator.LineValidator;
 import subway.domain.Line;
 import subway.domain.LineRepository;
