+++ src/test/java/subway/controller/validator/SectionValidatorTest.java
@@ -1,8 +1,10 @@
-package subway.controller.exception;
+package subway.controller.validator;
 
 import org.junit.Before;
 import org.junit.Test;
 
+import subway.controller.exception.IllegalElementException;
+import subway.controller.exception.NotExistedElementException;
 import subway.controller.validator.SectionValidator;
 import subway.domain.Line;
 import subway.domain.LineRepository;
