+++ src/test/java/subway/controller/validator/StationValidatorTest.java
@@ -1,8 +1,12 @@
-package subway.controller.exception;
+package subway.controller.validator;
 
 import org.junit.Test;
 
 import subway.controller.LineController;
+import subway.controller.exception.DuplicationException;
+import subway.controller.exception.IllegalElementException;
+import subway.controller.exception.NameFormatException;
+import subway.controller.exception.NotExistedElementException;
 import subway.controller.validator.StationValidator;
 import subway.domain.Station;
 import subway.domain.StationRepository;
